+++ src/Stormancer.Plugins/Firebase/Stormancer.server.Plugins.Firebase/FirebaseAuthenticationProvider.cs
@@ -63,7 +63,7 @@ namespace Stormancer.server.Plugins.Firebase
                 if (user == null)
                 {
                     var uid = Guid.NewGuid().ToString("N");
-                    user = await users.CreateUser(uid, JObject.FromObject(new { uid = firebaseUid }));
+                    user = await users.CreateUser(uid, JObject.FromObject(new { uid = firebaseUid }), Type);
                     user = await users.AddAuthentication(user, Type, claim => claim[ClaimPath] = firebaseUid, new Dictionary<string, string> { { ClaimPath, firebaseUid } });
                 }