+++ src/Stormancer.Plugins/Users/Stormancer.Server.Plugins.Users/Changelog.rst
@@ -14,6 +14,7 @@ Unreleased
 Fixed
 *****
 - Calling ConfigureUsers several times doesn't reset the config between two calls.
+- Removed IUSerService.CreateUser() overload with only 2 arguments: Providers should always specify the current platform when creating users.
 
 5.2.3.12
 --------