+++ src/Stormancer.Plugins/Users/Stormancer.Server.Plugins.Users/Changelog.rst
@@ -22,7 +22,7 @@ Added
 - AuthenticationProvider can add metadata to login results.
 Changed
 *******
-- Target dotnet 6 and Stormancer.Abstractions.Server 7.0.0
+- Target dotnet 6 and Stormancer.Abstractions.Server 7.1.0
 
 5.2.3.12
 --------